+++ indexer/scales.hpp
@@ -16,9 +16,9 @@ namespace scales
   int GetScaleLevel(m2::RectD const & r);
 
   /// @return such ration, that GetScaleLevel(ration) == level
-  double GetRationForLevel(int level);
+  double GetRationForLevel(double level);
   /// @return such rect, that GetScaleLevel(rect) == level
-  m2::RectD GetRectForLevel(int level, m2::PointD const & center, double X2YRatio);
+  m2::RectD GetRectForLevel(double level, m2::PointD const & center, double X2YRatio);
 
   double GetEpsilonForLevel(int level);
   double GetEpsilonForSimplify(int level);
